NativeFn is now comparable
This change changes the NativeFn type from a func to a *struct
containing a func. This allows functions to be compared and indexed.
E.g.
  print(map == map)  ==> true
  var m = { [map]: true }
  print(m[map])  ==> true

The downside is that "methods" became more inefficient as they now
require a memory allocation. But methods are currently a hack which
deserve to be implemented properly.