Package implementation - checkpoint
This change takes a step towards supporting packages (currently only
single unit scripts are supported).

This introduces the concept of a simple script as a special case of
package where the compiler wraps the script in a special `$main`
function.

I had to also create a new special $init function where globals are
initialized. At the moment, the only proper globals are meta parameters.